Add MFEMSuperLUSolver and remove deprecated hephaestus_solvers
