+++ src/problem_builders/problem_builder_base.C
@@ -103,98 +103,6 @@ ProblemBuilder::AddBoundaryCondition(std::string bc_name,
 }
 
 void
-ProblemBuilder::ConstructJacobianSolverWithOptions(SolverType type, SolverParams default_params)
-{
-  const auto & solver_options = GetProblem()->_solver_options;
-
-  const auto tolerance =
-      solver_options.GetOptionalParam<float>("Tolerance", default_params._tolerance);
-  const auto abs_tolerance =
-      solver_options.GetOptionalParam<float>("AbsTolerance", default_params._abs_tolerance);
-  const auto max_iter =
-      solver_options.GetOptionalParam<unsigned int>("MaxIter", default_params._max_iteration);
-  const auto print_level =
-      solver_options.GetOptionalParam<int>("PrintLevel", default_params._print_level);
-  const auto k_dim = solver_options.GetOptionalParam<unsigned int>("KDim", default_params._k_dim);
-
-  auto preconditioner =
-      std::dynamic_pointer_cast<mfem::HypreSolver>(GetProblem()->_jacobian_preconditioner);
-
-  switch (type)
-  {
-    case SolverType::HYPRE_PCG:
-    {
-      auto solver = std::make_shared<mfem::HyprePCG>(GetProblem()->_comm);
-
-      solver->SetTol(tolerance);
-      solver->SetAbsTol(abs_tolerance);
-      solver->SetMaxIter(max_iter);
-      solver->SetPrintLevel(print_level);
-
-      if (preconditioner)
-        solver->SetPreconditioner(*preconditioner);
-
-      GetProblem()->_jacobian_solver = solver;
-      break;
-    }
-    case SolverType::HYPRE_GMRES:
-    {
-      auto solver = std::make_shared<mfem::HypreGMRES>(GetProblem()->_comm);
-
-      solver->SetTol(tolerance);
-      solver->SetAbsTol(abs_tolerance);
-      solver->SetMaxIter(max_iter);
-      solver->SetKDim(k_dim);
-      solver->SetPrintLevel(print_level);
-
-      if (preconditioner)
-        solver->SetPreconditioner(*preconditioner);
-
-      GetProblem()->_jacobian_solver = solver;
-      break;
-    }
-    case SolverType::HYPRE_FGMRES:
-    {
-      auto solver = std::make_shared<mfem::HypreFGMRES>(GetProblem()->_comm);
-
-      solver->SetTol(tolerance);
-      solver->SetMaxIter(max_iter);
-      solver->SetKDim(k_dim);
-      solver->SetPrintLevel(print_level);
-
-      if (preconditioner)
-        solver->SetPreconditioner(*preconditioner);
-
-      GetProblem()->_jacobian_solver = solver;
-      break;
-    }
-    case SolverType::HYPRE_AMG:
-    {
-      auto solver = std::make_shared<mfem::HypreBoomerAMG>();
-
-      solver->SetTol(tolerance);
-      solver->SetMaxIter(max_iter);
-      solver->SetPrintLevel(print_level);
-
-      GetProblem()->_jacobian_solver = solver;
-      break;
-    }
-    case SolverType::SUPER_LU:
-    {
-      auto solver = std::make_shared<platypus::SuperLUSolver>(GetProblem()->_comm);
-
-      GetProblem()->_jacobian_solver = solver;
-      break;
-    }
-    default:
-    {
-      MFEM_ABORT("Unsupported solver type specified.");
-      break;
-    }
-  }
-}
-
-void
 ProblemBuilder::ConstructNonlinearSolver()
 {
   auto nl_solver = std::make_shared<mfem::NewtonSolver>(GetProblem()->_comm);
